+++ CHANGELOG.md
@@ -17,9 +17,15 @@
     Click to see more.
   </summary>
 
+
+### :bug: Bug Fix
+- `video`
+  - [#3682](https://github.com/wix/ricos/pull/3682) fix video duration on tiptap and converter
+
 ### :nail_care: Polish
  - `audio`
   - [#3700](https://github.com/wix/ricos/pull/3700) change TextInput doubleClick prop name 
+
 
 </details>
   <hr/>